Small project structure re-organization before next major step.
