+++ Alpaca.Markets/Interfaces/IAsset.cs
