+++ Alpaca.Markets/Interfaces/ICalendar.cs
