+++ Alpaca.Markets/Interfaces/IClock.cs
